# cut: add whitespace option for separating fields