--- a/src/uu/cut/src/cut.rs
+++ b/src/uu/cut/src/cut.rs
@@ -264,7 +264,6 @@ fn cut_fields_delimiter<R: Read>(
     Ok(())
 }
 
-#[allow(clippy::cognitive_complexity)]
 fn cut_fields_whitespace<R: Read>(
     reader: R,
     ranges: &[Range],