… docs/src/extensions.md
… tests/by-util/test_cut.rs
… tests/fixtures/cut/whitespace_delimited.expected