**TLDR**: "-w" flag in cut is an extension option implemented in [FreeBSD](https://www.freebsd.org/cgi/man.cgi?cut). This option proves to be useful when cutting fields separated by one or more whitespaces. GNU coreutils does not currently implement this option. This PR implements -w option and adds test cases. The performance difference compared to FreeBSD's cut implementation is about 9x.

**Example use case**: "ls" or "ps" returns columns separated by multiple inconsistent whitespaces. We can use -w option to select columns, e.g. "ps | cut -w -f1" to print only pids.

**Implementation**: For best performance, I created a new function (cut_fields_whitespace) dedicated for -w option. Performance of "cut -f" before and after this PR is essentially the same, as show below:

Before
```
$ hyperfine -w3 "./target/release/coreutils cut -f2-4,8 data.tsv"
Benchmark 1: ./target/release/coreutils cut -f2-4,8 data.tsv
  Time (mean ± σ):      1.126 s ±  0.006 s    [User: 1.035 s, System: 0.090 s]
  Range (min … max):    1.118 s …  1.138 s    10 runs

$  hyperfine -w3 "./target/release/coreutils cut -f2-4,8 -d' ' data.tsv"
Benchmark 1: ./target/release/coreutils cut -f2-4,8 -d' ' data.tsv
  Time (mean ± σ):      1.026 s ±  0.003 s    [User: 0.940 s, System: 0.086 s]
  Range (min … max):    1.023 s …  1.031 s    10 runs
```

After
```
$ hyperfine -w3 "./target/release/coreutils cut -f2-4,8 data.tsv"
Benchmark 1: ./target/release/coreutils cut -f2-4,8 data.tsv
  Time (mean ± σ):      1.127 s ±  0.021 s    [User: 1.037 s, System: 0.089 s]
  Range (min … max):    1.110 s …  1.169 s    10 runs

$ hyperfine -w3 "./target/release/coreutils cut -f2-4,8 -d' ' data.tsv"
Benchmark 1: ./target/release/coreutils cut -f2-4,8 -d' ' data.tsv
  Time (mean ± σ):      1.045 s ±  0.026 s    [User: 0.956 s, System: 0.088 s]
  Range (min … max):    1.025 s …  1.103 s    10 runs
```

The performance of "cut -w -f" compared to FreeBSD is about 9x faster:
```
$ hyperfine -w3 "./target/release/coreutils cut -f2-4,8 -w data.tsv" "/usr/bin/cut -f2-4,8 -w data.tsv"
Benchmark 1: ./target/release/coreutils cut -f2-4,8 -w data.tsv
  Time (mean ± σ):      1.242 s ±  0.003 s    [User: 1.153 s, System: 0.088 s]
  Range (min … max):    1.240 s …  1.249 s    10 runs

Benchmark 2: /usr/bin/cut -f2-4,8 -w data.tsv
  Time (mean ± σ):     11.529 s ±  0.089 s    [User: 11.385 s, System: 0.131 s]
  Range (min … max):   11.451 s … 11.731 s    10 runs

Summary
  './target/release/coreutils cut -f2-4,8 -w data.tsv' ran
    9.28 ± 0.08 times faster than '/usr/bin/cut -f2-4,8 -w data.tsv'
```