@tertsdiepraam 2022-12-13
I think they're here: https://github.com/freebsd/freebsd-src/tree/373ffc62c158e52cde86a5b934ab4a51307f9f2e/usr.bin/cut/tests

Though this is not very extensive.