@TechHara 2022-12-14
> I think they're here: https://github.com/freebsd/freebsd-src/tree/373ffc62c158e52cde86a5b934ab4a51307f9f2e/usr.bin/cut/tests
> 
> Though this is not very extensive.

and some here too, but not much
https://github.com/freebsd/freebsd-src/tree/373ffc62c158e52cde86a5b934ab4a51307f9f2e/contrib/netbsd-tests/usr.bin/cut