@TechHara 2022-12-18
> GNU testsuite comparison:
> 
> ```
> GNU test failed: tests/tail-2/inotify-dir-recreate. tests/tail-2/inotify-dir-recreate is passing on 'main'. Maybe you have to rebase?
> ```

This is not related to this PR and seems to be failing in other PRs as well