@TechHara 2022-12-18
> for multispace, add it to the top of the file in the ignore spell list :)

done!