@tertsdiepraam 2023-01-01
Yeah I think it's good enough now as far as adding the feature goes. I do think `cut` is due for a refactor that removes some of the duplicated code though.